PR30305: Implement proposed DR resolution to prevent slicing via inherited constructor.
The rule we use is that a construction of a class type T from an argument of
type U cannot use an inherited constructor if U is the same as T or is derived
from T (or if the initialization would first convert it to such a type). This
(approximately) matches the rule in use by GCC, and matches the current proposed
DR resolution.


git-svn-id: http://llvm.org/svn/llvm-project/cfe/trunk@291403 91177308-0d34-0410-b5e6-96231b3b80d8